+++ Editor/StateMachines/ShapeState_DrawHeightShape.cs
@@ -61,7 +61,8 @@ namespace UnityEditor.ProBuilder
                 }
             }
 
-            tool.DrawBoundingBox();
+            if(evt.type == EventType.Repaint)
+                tool.DrawBoundingBox();
 
             if(evt.isMouse)
             {